+++ data-raw/barplot_MinSta.R
@@ -1,21 +1,22 @@
 ## code to prepare `barplot_table_plot_pattern` dataset goes here
 barplot_table_plot_pattern <- plotsettings_tablebarplot(
-  background_stripes_colour = c(rep(c("white", "white", "white", "gray94", "gray94", "gray94"), 8), rep("gray87", 3)),
+  background_stripes_colour = c(rep(c("white", "white", "white", rep(cmyk(7, 0, 4, 1), 3)), 8), rep(cmyk(0, 0, 0, 10), 3)),
   bar_background_lines = "border",
   bar_background_lines_linetype = "dotted",
   bar_fill_colour = c(
-    grDevices::rgb(33, 89, 104, maxColorValue = 255),
-    grDevices::rgb(49, 133, 156, maxColorValue = 255),
-    grDevices::rgb(75, 172, 198, maxColorValue = 255)
+   cmyk(85, 0, 43, 17),
+   cmyk(40, 0, 20, 8),
+   cmyk(20, 0, 10, 4)
   ),
   bar_frame_linetype = c(`TRUE` = "solid", `FALSE` = "dashed"),
+  bar_line_width = 0.1,
   bar_pattern_fill_colour = "white",
   bar_pattern_type = c("FALSE" = "stripe", "TRUE" = "none"),
   bar_sig_type = "pattern",
   bar_width = 0.7,
   columns_table_sig_high_letter = "b",
-  columns_table_sig_high_letter_nudge_x = 0.04,
-  headers_background_colour = "lightblue",
+  columns_table_sig_high_letter_nudge_x = 4.7,
+  headers_background_colour = cmyk(40, 0, 20, 8),
   headers_background_width_y = 0.5,
   font_size = 2,
   bar_pattern_spacing = 0.01,
@@ -26,9 +27,9 @@ usethis::use_data(barplot_table_plot_pattern, overwrite = TRUE)
 
 
 barplot_plot_frame <- plotsettings_tablebarplot(
-  background_stripes_colour = c(rep(c("white", "gray94"), 8), "gray87"),
+  background_stripes_colour = c(rep(c("white"), 16), cmyk(0, 0, 0, 15)),
   bar_background_lines = "none",
-  bar_fill_colour = c("lightblue"),
+  bar_fill_colour = cmyk(40, 0, 20, 8),
   bar_frame_linetype = c("above" = "solid", "below" = "longdash", "no_sig" = "blank"),
   bar_label_nudge_x = -0.2,
   bar_label_size = 1.75,